requirements.txt
# Core dependencies
supabase==2.3.0
python-dotenv==1.0.0
requests==2.31.0

# PDF processing libraries
PyMuPDF==1.23.8
pdfplumber==0.10.3

# Data processing
pandas==2.1.4
numpy==1.24.3